fix: Correct VizJS integration in unified template
- Fix PlantUML and Graphviz rendering by properly integrating VizJS
- Remove problematic script tag embedding in innerHTML
- VizJS now executes properly in renderDiagram function
- Create svg-output div and populate with VizJS-generated SVG
- All three renderers now work correctly with unified template
- Consistent error handling across all diagram types
